🎨 style(globals.scss): decrease h3 font size to 1.8rem 🔧 chore(Contact/index.js): change API endpoint to localhost for development purposes 🔧 chore(User/index.js): change API endpoint to localhost for development purposes 🔧 chore(getArticleList.js): change API endpoint to localhost for development purposes The h3 font size was decreased to 1.8rem for better visual consistency. The API endpoint for Contact, User, and getArticleList components were changed to localhost for development purposes.
🔧 chore(getDeliveryArea.js, getProductList.js, getReviewList.js): switch API endpoint to localhost
The API endpoint has been switched to localhost for the getDeliveryArea, getProductList, and getReviewList modules. The previous endpoint was commented out and replaced with the localhost endpoint. This change was made to allow for local development and testing.

🔥 chore(User/index.js, index.js): remove unused reducers
The User reducer and the index.js file that imports and combines the reducers have been removed. These reducers were not being used in the application and were therefore removed to improve code cleanliness and reduce clutter.

🔥 chore(settings): remove settings reducer
The settings reducer has been removed from the project. This commit removes the file that contained the reducer and all of its associated code.

🔥 chore(store.js): remove unused store.js file
The store.js file was removed as it was no longer being used in the application.